ias_ParametrisationUpdate: many changes
Working now with a Neohookean energy and a constraint for normal
motion. The constraint is imposed with a Lag Mult (local pressure)
To avoid inf-sup, a small diagonal matrix is added to the Hessian
(not modifying the residual but effectively giving stability to the
solution). Many comments are left in case parts of the code are needed.
Next commit will erase these comments.