MAJOR CHANGE:  calculating interacting Gauss points now in Integration
This affects all Tutorials.

Previously, we calculated interacting elements in a very basic way in
the Tissue object. Now, instead, we calculate which Gauss points interact
with the others in the opposite mesh, using for this the Vtk function
FindPointsWithinRadius. This is faster and consistent with the integration,
which maybe was not the case previously.