MAJOR CHANGE: interacting Gauss points are now calculated every NR iteration
This can be enabled or disabled.

The cutoff distance is now set in the integration object and used
in subsequent calls to calculateInteractingGaussPoints. If one sets
the displacement fields in the integration with setDisplacementFieldNames
then these displacements are added to the calculation of the interacting
Gauss points (so it is consistent).

The user can ask whether the matrix structure needs to be recalculated with
getRecalculateMatrixStructure also in the Integration object.